+++ jina/types/arrays/mixins/magic.py
@@ -1,6 +1,7 @@
 from typing import TYPE_CHECKING
 
 from ....helper import typename
+from ....proto import jina_pb2
 
 if TYPE_CHECKING:
     from ..types import DocumentArraySourceType
@@ -31,3 +32,26 @@ class MagicMixin:
         for doc in other:
             self.append(doc)
         return self
+
+    def __getstate__(self):
+        dap = jina_pb2.DocumentArrayProto()
+        dap.docs.extend(self._pb_body)
+        r = dict(serialized=dap.SerializePartialToString())
+        if hasattr(self, '_ref_doc'):
+            r['ref_doc'] = self._ref_doc.binary_str()
+        return r
+
+    def __setstate__(self, state):
+        dap = jina_pb2.DocumentArrayProto()
+        dap.ParseFromString(state['serialized'])
+        from ..document import DocumentArray
+
+        da = DocumentArray(dap.docs)
+        if 'ref_doc' in state:
+            from ...document import Document
+
+            ref_doc = Document(state['ref_doc'])
+            self.__init__(da, ref_doc)
+        else:
+            self.__init__()
+            self.extend(da)