+++ jina/types/document/mixins/sugar.py
@@ -6,6 +6,7 @@ if TYPE_CHECKING:
     from ...arrays.memmap import DocumentArrayMemmap
     from ...ndarray import ArrayType
     from ...arrays.mixins.embed import AnyDNN
+    from ....helper import T
 
     import numpy as np
 
@@ -15,7 +16,7 @@ class SingletonSugarMixin:
 
     @overload
     def match(
-        self,
+        self: 'T',
         darray: Union['DocumentArray', 'DocumentArrayMemmap'],
         metric: Union[
             str, Callable[['ArrayType', 'ArrayType'], 'np.ndarray']
@@ -27,7 +28,8 @@ class SingletonSugarMixin:
         exclude_self: bool = False,
         only_id: bool = False,
         use_scipy: bool = False,
-    ) -> None:
+        num_worker: Optional[int] = None,
+    ) -> 'T':
         """Matching the current Document against a set of Documents.
 
         The result will be stored in :attr:`.matches`.
@@ -44,34 +46,39 @@ class SingletonSugarMixin:
                                 the min distance will be rescaled to `a`, the max distance will be rescaled to `b`
                                 all values will be rescaled into range `[a, b]`.
         :param metric_name: if provided, then match result will be marked with this string.
-        :param batch_size: if provided, then `darray` is loaded in chunks of, at most, batch_size elements. This option
-                           will be slower but more memory efficient. Specialy indicated if `darray` is a big
-                           DocumentArrayMemmap.
+        :param batch_size: if provided, then ``darray`` is loaded in batches, where each of them is at most ``batch_size``
+            elements. When `darray` is big, this can significantly speedup the computation.
         :param exclude_self: if set, Documents in ``darray`` with same ``id`` as the left-hand values will not be
                         considered as matches.
         :param only_id: if set, then returning matches will only contain ``id``
         :param use_scipy: if set, use ``scipy`` as the computation backend. Note, ``scipy`` does not support distance
             on sparse matrix.
+        :param num_worker: the number of parallel workers. If not given, then the number of CPUs in the system will be used.
+
+                .. note::
+                    This argument is only effective when ``batch_size`` is set.
         """
         ...
 
-    def match(self, *args, **kwargs) -> None:
+    def match(self: 'T', *args, **kwargs) -> 'T':
         """
         # noqa: D102
         # noqa: DAR101
+        :return: itself after modified
         """
         from ...arrays import DocumentArray
 
         _tmp = DocumentArray([self])
         _tmp.match(*args, **kwargs)
+        return self
 
     @overload
     def embed(
-        self,
+        self: 'T',
         embed_model: 'AnyDNN',
         device: str = 'cpu',
         batch_size: int = 256,
-    ) -> None:
+    ) -> 'T':
         """Fill the embedding of Documents inplace by using `embed_model`
 
         :param embed_model: the embedding model written in Keras/Pytorch/Paddle
@@ -80,12 +87,14 @@ class SingletonSugarMixin:
         :param batch_size: number of Documents in a batch for embedding
         """
 
-    def embed(self, *args, **kwargs) -> None:
+    def embed(self: 'T', *args, **kwargs) -> 'T':
         """
         # noqa: D102
         # noqa: DAR101
+        :return: itself after modified.
         """
         from ...arrays import DocumentArray
 
         _tmp = DocumentArray([self])
         _tmp.embed(*args, **kwargs)
+        return self